Modified createHmatrix functions
Modified createHmatrix inner functions so that the size of the H matrix
is the number of local measurements x the length of each local state
vector. Modified realPowerInjMeas and reactivePowerInjMeas so that they
only use local variables, including local G and B matrices. Not 100%
sure if they're implemented correctly, especially for the sum across all
buses. Now it's only accounting for buses that are connected to the bus
of interest.