Should try adding a ton of measurements to the 14 bus ADMM system
I tried adding a few select measurements to Partition 1 in
example_14bus_IEEE_rectADMM, but they didn't seem to help the ADMM
convergence (in fact, made it worse). I also coded up an automatic way
to aggregate c_k, but it complicates how numPart is calculated... It
might be better to go back to yesterday's commit as a starting point.
The next thing to try is to probably add a ton more measurements to the
whole system.