How to reference areas with different slack buses to the same global slack bus
One way to reference all of the areas to the same slack bus is to look
at common buses measured between the area of interest and the area with
the global slack bus. Then take the angle(s) of the area with the global
slack bus and subtract from it the angle(s) of the area of interest.
Then take the average of all the angles and add that to the area of
interest's state vector.