Partially fixed the slack bus shift issue
New code to add the calculated slack bus difference to the state vectors
of Areas 2-4. The problem now though is that when you add the slack bus
difference, the initial f_slack for that area goes from 0 to an actual
value. However, all the state vectors assume that the length is
2*(numbus of area) - 1. Need to reconcile that difference. I'm not sure
that's the only issue either, because when it actually compiled without
errors and ran, it still didn't converge. I suspect that there's an off
by 1 index issue due to the unremoved slack f entry.