Checked how dx_k is calculated.
Checking to see whether adding an extra zero column to H is going to
change the solution. The answer is, the Gain is exactly the same with an
extra row and column of 1s for the slack bus. g1 is the same with an
extra zero in the slack bus row. The effect of -Gain\g1 is almost
negligible with the solution being different by max of 2e-5.