2 partition 118 case now runs, but the residuals are completely wrong.
Identified the 118 bus tie lies and slightly edited the way the auto
measurements work. For example, separated out regular buses and overlap
tie buses. However, it's still not quite working and very hard to debug
the 118 case, so go back to the 14 bus case and run the auto measurement
code to see if it works there. So far I'm not sure, because there's
something wrong with the way PowerWorld is writing the Ybus file. I keep
getting a weird I/O error. Also, I need to implement the double lines
feature.