Added double lines feature to the centralized 14 bus and 118 bus cases, trying to debug 118 ADMM case
I implemented the double lines feature in the 14 bus and 118 bus cases.
When you compare the 118 results against the PW power flow results,
remember that for some reason PowerWorld sometimes shifts the angles. So
in the 118 bus case, the AC slack bus angle is NOT equal to 0. When you
account for that, the two results are the same. However, now I'm trying
to figure out why the 118 ADMM case doesn't converge. There are a number
of errors in the P,Q results for a number of buses. Also bus 59 isn't
included in either partition. Need to fix these.