ADMM 118 bus case now works.
It works!!! Now converges in 5 iterations with an accuracy of 1e-6.
Debugged both h vectors and H matrices (particularly H2). The main
issues were 1) slack2 was set to bus 1 instead of bus 69, 2) bus 59 was
not included in either partition, 3) bus 58? was somehow in both
partitions, and 4) the singular matrix error was from me accidentally
commenting out the line in createHmatrix_ADMM to refill in the slack bus
column with zeros.