Used METIS to generate partitions of 118 case. Also changed getPartitions to auto find tie lines.
Used METIS to generate partitions of 2, 4, 8, and 16 for the 118 case.
Also changed getPartitions so that it can automatically figure out what
the tie lines should be for each partition. Need to verify that it's
actually working correctly. Also continue to tie together code so that I
could time serially how long each partition's SE calculation and
communication take.