Issue with auto partitions converging for ADMM, non-contiguous METIS partitions
1. Used METIS to break up 118 case into 3 and 4 partitions. Traced
through PW one line and saw that for 2 partitions, METIS says they're
contiguous, but they're not actually contiguous (see IEEE 118
Bus_METIS2parts.pwd). Could that be a reason for not converging ADMM?
2. There is currently an issue getting the auto partitions to converge
in ADMM. It also had a similar singular issue for the centralized one,
but it turns out to be because I had initially specified numParts = 1,
but I had to specify numParts = 2. Otherwise, there were empty columns
in H, thus making the Gain matrix singular. Try spy on ADMM H's
tomorrow, and see if there are empty columns.