Preliminary timing results in the .xlsx file; automatically identify the global slack area
The auto partitions are still not converging for ADMM. I tried spy on
Hs, but they both appeared to be full rank. I checked the H matrices
against the centralized one, and they appear to be the same except for
the slack bus. My new guess is that somehow the slack bus selection is
messing things up.

The following things are in this commit:
1. I performed some preliminary timing tests for centralized and
manually partitioned 118 (tic toc outside both while loops).
Surprisingly the ADMM serial one is (~2x) faster than the centralized
one, even though there's more complexity in the formulation.
2. I think there's an issue with auto reconciling the slack buses. I
added a little thing to automatically identify the global slack bus
(i.e. = bus 1 always since that's how the centralized one is set up).

To do:
Based on the auto global slack area, I need to make sure that the
addSlack function is calculating correctly.