Fixed auto partitioning ADMM convergence issue.
As I suspected, the issue had something to do with the slack bus.
Basically when I originally coded everything, I assumed that the global
slack area is always 1, but in reality it should be whatever area
contains the global slack bus, currently set to be bus 1. Hence, when
you calculate the addSlack, it always references it to area 1. However,
in my auto partitions from METIS, it can be any area.

To do for next time:
Right now the working code is in test_14busADMM_partitions. This should
be cleaned up so that either the 14 bus or the 118 bus can run in the
same script. Make sure that it works properly for the 118 case. Then
extend to bigger cases, and time the result. Also, it would be
interesting to compare the results across the auto partitions and the
hand partitions to see if there's a speed up.