Fixed the issue that caused warning about Gain being close to singular/badly scaled
MATLAB gave a warning at every iteration for Partition #2 in an auto
2-area partition about the gain matrix being close to singular/badly
scaled. It turns out that if I don't force METIS to give me contiguous
partitions, then part of the control area is unobservable and thus the
matrix is close to singular. Fixed the issue by forcing METIS to give me
contiguous partitions, which I'm not sure how well METIS can enforce.
When I ran that again for the auto 2 partition, the answer converged to
the central solution.

However, for the auto 4 partition, now there is a new issue where the
normres_r does not approach 0, due to the answers from Partition 1 being
off. Why is that?