Found reason why normres_r doesn't converge to 0
I found the reason why the normres_r doesn't converge to 0. I confirmed
on the 118 for 3 partitions. Basically the slack area wasn't connected
to one of the other partitions, and thus that partition was not
referenced to the global slack bus. The solution I'm thinking of now is
to create an area adjacency matrix, use Dijkstra's algorithm to find the
shortest path, and then chain reference back to the global slack area.