Finally got timing data for the 118 case for numParts = 2 to 32, see .xlsx
The overall algorithm finally works. However, I coded up the
ref2GlobalSlack function as a centralized BFS. In reality, it would be
much more decentralized and thus take longer. Some analysis on this is
needed. An alternate solution would be to simply use PMUs. For now, it
looks like for the 118 case, 10 areas is about optimal. However, even
for 32 areas, it's not that much slower than the optimal case. Also it's
still much much better than the centralized case.