TVA case still doesn't converge
The gain matrices for Areas 14, 17, and 19 are singular. One possibility
is that I removed the triple and quadruple lines in the TVA case, thus
eliminating bus 1308 and impacting the measurements for buses connected
to 1308. Another difference between the TVA case and the other cases I
tried is the fact that it has multiple lines with 3-4 parallel branches.
In the 14 bus and 118 bus cases, those only had a max of 2 parallel
branches. The other differences are trying to account for the tap ratios
(make them fixed?) and possibly phase shifter.

Go through each one of these issues for the 14 bus case, and see if that
could potentially be it.