Lots of changes due to revert; prelim data for GM 2016
A lot of changes in this commit, because I freaked out and reverted to
an older version when I thought the 14 bus system wasn't working. It is
now, and I have some preliminary data for that plus the 118. Next time,
try to get the 57 bus system working and see if can draw a trend from
the number of partitions vs computational speed.