Fixed ill-conditioned Gain, also fixed convergence issue from non-contiguous partitions
Fixed ill-conditioned Gain issue as described in the last commit.
After that, partition 8 was still causing issues and not converging. It
turns out it was because Partition 8 was non-contiguous. Can enforce
contiguousness using METIS k-way only. Really need to create a
contiguity checker and move buses to other partitions if it's not
contiguous.