Started adding Gaussian noise to measurements
Generated a "good" noise signal with mean 0 and std 0.01. Not adding
Gaussian noise to measurements for the global slack bus (possibly also
local slack buses), which might complicate issues. Need to write
function that maps the global measurement indices to the local indices,
so that both sets are seeing the exact same measurements.