Vectorized h, which makes central SE run 10x faster
Vectorized the way h runs, which is now much faster. For example, now
TVA only takes ~5 min instead of ~30 min.

Got rid of one layer of for loops in how H is calculated. Still need to
vectorize each type of function. That is the majority of the remaining
time sink.

TVA case still has small errors, not sure why.